implementing the filter system

+++ lightreporter/src/main/scala/com/lightreporter/Filter/opt/optImpl/IntOpt.scala
@@ -7,7 +7,7 @@ import com.lightreporter.Filter.opt.{ComparableOpt, Operator, ValueGetter}
   */
 class IntOpt[T](optName:String, value:String, valueSelector: ValueGetter[T]) extends ComparableOpt[T,Int](optName,value,valueSelector){
 
-  val intVal=value.toInt
+  val intVal=value.toDouble.toInt
 
   override def isPass(msg: T): Boolean = {
     val msgBool=valueSelector.getVal(msg).asInstanceOf[Int]
