+++ lightreporter/src/main/scala/com/lightreporter/Filter/opt/optImpl/LongOpt.scala
@@ -7,7 +7,7 @@ import com.lightreporter.Filter.opt.{ComparableOpt, Operator, ValueGetter}
   */
 class LongOpt[T](optName:String, value:String, valueSelector: ValueGetter[T]) extends ComparableOpt[T,Long](optName,value,valueSelector){
 
-  val longVal=value.toLong
+  val longVal=value.toDouble.toLong
 
   override def isPass(msg: T): Boolean = {
     val msgBool=valueSelector.getVal(msg).asInstanceOf[Long]
