+++ lightreporter/src/main/scala/com/lightreporter/Filter/opt/optImpl/ShortOpt.scala
@@ -7,7 +7,7 @@ import com.lightreporter.Filter.opt.{ComparableOpt, Operator, ValueGetter}
   */
 class ShortOpt[T](optName: String, value: String, valueSelector: ValueGetter[T]) extends ComparableOpt[T, Short](optName, value, valueSelector) {
 
-  val shortVal = value.toShort
+  val shortVal = value.toDouble.toShort
 
   override def isPass(msg: T): Boolean = {
     val msgBool = valueSelector.getVal(msg).asInstanceOf[Short]
